chore(api) move db helper functions & migrations into new 'database' dir
